Now avoids recloning the repo every start

+++ entrypoint.sh
@@ -6,24 +6,31 @@ if [ -z "$SSH_KEY" ]; then
   exit 1
 fi
 
-# Copy the SSH key from the environment variable
-echo "${SSH_KEY}" > ~/.ssh/id_rsa
-chmod 600 ~/.ssh/id_rsa
 
-# Start SSH agent and add the key
-eval $(ssh-agent -s)
-ssh-add ~/.ssh/id_rsa
+if [ ! -f "/root/.ssh/id_rsa" ]; then
+  # Copy the SSH key from the environment variable
+  echo "${SSH_KEY}" > ~/.ssh/id_rsa
+  chmod 600 ~/.ssh/id_rsa
 
-# Add GitHub to known hosts to avoid interactive prompt
-ssh-keyscan -H github.com >> ~/.ssh/known_hosts
+  # Start SSH agent and add the key
+  eval $(ssh-agent -s)
+  ssh-add ~/.ssh/id_rsa
 
-# Clone the repository
-mkdir -p /work/temp
-git clone git@github.com:DanielGardin/auto-bidding.git /work/temp
-rsync -a /work/temp/* /work/auto-bidding
-rm -rf /work/temp
+  # Add GitHub to known hosts to avoid interactive prompt
+  ssh-keyscan -H github.com >> ~/.ssh/known_hosts
+fi
 
-export PYTHONPATH="/root/biddingTrainEnv:${PYTHONPATH}"
+
+if [ ! -d "/work/auto-bidding/.git" ]; then
+  # Clone the repository
+  mkdir -p /work/temp
+  git clone git@github.com:DanielGardin/auto-bidding.git /work/temp
+  shopt -s dotglob
+  rsync -a /work/temp/* /work/auto-bidding
+  rm -rf /work/temp
+
+  export PYTHONPATH="/root/biddingTrainEnv:${PYTHONPATH}"
+fi
 
 # Run the main process
 exec "$@"
